P3-Behavioral-Cloning-XiaoCar/requirements.txt
flask==0.12
# requires sudo apt-get install libi2c-dev i2c-tools python3-dev libffi-dev
smbus-cffi==0.5.1
scipy
sklearn
keras
#tensorflow>=0.12 for installation on Pi board see tensoflow-install-pi.sh
apscheduler
